removed model and updated a bunch of stuff, should be report version
