+++ .gitignore
@@ -15,3 +15,4 @@ __pycache__
 __pycache__/
 */__pycache__
 models/__pycache__/
+PointNetModel.keras
