+++ python_scripts/jets_training/jets_train.py
@@ -13,11 +13,7 @@ from models.JetPointNet import PointNetSegmentation, masked_bce_loss, masked_mae
 
 os.environ['CUDA_VISIBLE_DEVICES'] = "4" # SET GPU
 
-
-
 MAX_SAMPLE_LENGTH=859
-
-
 
 class BatchNormalizationMomentumScheduler(tf.keras.callbacks.Callback):
     def __init__(self, initial_momentum=0.5, final_momentum=0.99, total_epochs=20):
@@ -112,7 +108,7 @@ print("Training on Dataset: ", TRAIN_DIR)
 lr_scheduler = tf.keras.callbacks.LearningRateScheduler(scheduler)
 csv_logger = CSVLogger('saved_model/training_log.csv', append=True, separator=';')
 save_model_callback = LambdaCallback(on_epoch_end=save_model_on_epoch_end)
-batch_norm_scheduler = BatchNormalizationMomentumScheduler(total_epochs=EPOCHS)
+# batch_norm_scheduler = BatchNormalizationMomentumScheduler(total_epochs=EPOCHS) # Was used in the pointnet paper, can decide if you'd like to
 
 
 start_time = time.time()
@@ -121,7 +117,7 @@ model.fit(train_generator,
           epochs=EPOCHS,
           validation_data=val_generator,
           validation_steps=val_steps,
-          callbacks=[csv_logger, save_model_callback, lr_scheduler, batch_norm_scheduler])  # Include lr_scheduler in callbacks
+          callbacks=[csv_logger, save_model_callback, lr_scheduler])  # Include lr_scheduler in callbacks
 end_time = time.time()
 
 print(f"Training Done! Took {(end_time - start_time) / 60 / 60} hours!")
