+++ python_scripts/jets_training/models/JetPointNet.py
@@ -3,7 +3,7 @@
 Adapted From
 https://github.com/lattice-ai/pointnet/tree/master
 
-Original Architecture From
+Original Architecture From Pointnet Paper:
 https://arxiv.org/pdf/1612.00593.pdf
 '''
 
@@ -12,6 +12,8 @@ import tensorflow as tf
 import numpy as np
 import keras 
 
+# =======================================================================================================================
+# ============ Weird Stuff ==============================================================================================
 
 
 class SaveModel(keras.callbacks.Callback):
@@ -19,11 +21,12 @@ class SaveModel(keras.callbacks.Callback):
         self.model.save("JetPointNet_{epoch}.hd5".format(epoch))
 
 class CustomMaskingLayer(tf.keras.layers.Layer):
+    # For masking out the inputs properly, based on points for which the last value in the point's array (it's "type") is "-1"
     def __init__(self, **kwargs):
         super(CustomMaskingLayer, self).__init__(**kwargs)
     
     def call(self, inputs):
-        mask = tf.not_equal(inputs[:, :, -1], -1)
+        mask = tf.not_equal(inputs[:, :, -1], -1) # Masking
         mask = tf.cast(mask, tf.float32)
         mask = tf.expand_dims(mask, -1)
         return inputs * mask
@@ -32,6 +35,7 @@ class CustomMaskingLayer(tf.keras.layers.Layer):
         return input_shape
 
 class OrthogonalRegularizer(tf.keras.regularizers.Regularizer):
+    # Used in Tnet in PointNet for transforming everything to same space
     def __init__(self, num_features=6, l2=0.001):
         self.num_features = num_features
         self.l2 = l2
@@ -48,10 +52,26 @@ class OrthogonalRegularizer(tf.keras.regularizers.Regularizer):
         return {'num_features': self.num_features, 'l2': self.l2}
     
 
+def TSSR_Activation(x):
+    # An activation function that's linear when |x| < 1 and (an adjusted) sqrt when |x| > 1 (is kinda like a "soft tanh" function that isn't bounded in output)
+    # Adapted from https://arxiv.org/pdf/2308.04832.pdf
+    # Careful, can produce negative values (can mess around with this, I think it might be good to stabilize training) - ie the problem of hard constraints vs soft constraints in an optimization problem
+    high_value_slope_coefficient = 1 #Value to multiply by the sqrt|x| part, lower means flatter, higher means steeper
+    condition = tf.abs(x) < 1
+    return tf.where(condition, x, tf.sign(x) * (high_value_slope_coefficient*tf.sqrt(tf.abs(x)) - high_value_slope_coefficient + 1))
+
+# =======================================================================================================================
+# =======================================================================================================================
+
+
+
+# =======================================================================================================================
+# ============ Main Model Blocks ========================================================================================
+
 def conv_mlp(input_tensor, filters, dropout_rate = None):
     # Apply shared MLPs which are equivalent to 1D convolutions with kernel size 1
     x = tf.keras.layers.Conv1D(filters=filters, kernel_size=1, activation='relu')(input_tensor)
-    x = tf.keras.layers.BatchNormalization(momentum=0.0)(x)
+    x = tf.keras.layers.BatchNormalization()(x)
     if dropout_rate is not None:
         x = tf.keras.layers.Dropout(dropout_rate)(x)
     return x
@@ -80,15 +100,12 @@ def TNet(input_tensor, size, add_regularization=False):
     x = tf.reshape(x, (-1, size, size))
     return x        
 
-def TSSR_Activation(x):
-    # From https://arxiv.org/pdf/2308.04832.pdf
-    condition = tf.abs(x) < 1
-    return tf.where(condition, x, tf.sign(x) * (2*tf.sqrt(tf.abs(x)) - 1))
 
 def PointNetSegmentation(num_points, num_classes):
     num_features = 6  # Number of input features
 
     network_size_factor = 5 # Mess around with this along with the different layer sizes
+
     input_points = tf.keras.Input(shape=(num_points, num_features))
 
     # T-Net for input transformation
@@ -105,6 +122,7 @@ def PointNetSegmentation(num_points, num_classes):
     x = conv_mlp(x, 128 * network_size_factor)
     x = conv_mlp(x, 1024 * network_size_factor)
 
+    # Get global features and expand
     global_feature = tf.keras.layers.GlobalMaxPooling1D()(x)
     global_feature_expanded = tf.keras.layers.Lambda(lambda x: tf.expand_dims(x, 1))(global_feature)
     global_feature_expanded = tf.keras.layers.Lambda(lambda x: tf.tile(x, [1, num_points, 1]))(global_feature_expanded)
@@ -117,16 +135,20 @@ def PointNetSegmentation(num_points, num_classes):
 
     # Extract energy from input and multiply by the segmentation output
     energy = tf.expand_dims(input_points[:, :, 4], -1)  # Assuming energy is at index 4
-    segmentation_output_pre_sigmoid = tf.keras.layers.Conv1D(num_classes, kernel_size=1)(c)  # No activation yet
-    segmentation_output_pre_sigmoid = tf.keras.layers.Activation(TSSR_Activation)(segmentation_output_pre_sigmoid)  # Apply sigmoid
+    segmentation_output_pre_sigmoid = tf.keras.layers.Conv1D(num_classes, kernel_size=1)(c)  # No activation yet ("sigmoid" here is a misnomer, we're using TSSR. Feel free to update)
+    segmentation_output_pre_sigmoid = tf.keras.layers.Activation(TSSR_Activation)(segmentation_output_pre_sigmoid) # Apply activation + adjust float back to 32 bit for training (a smarter way to do this probably exists)
     segmentation_output = tf.multiply(segmentation_output_pre_sigmoid, energy)  # Multiply by energy
-
-    #segmentation_output = tf.keras.layers.Conv1D(num_classes, kernel_size=1, activation = "relu")(c)  # No activation yet
 
     model = tf.keras.Model(inputs=input_points, outputs=segmentation_output)
 
     return model
 
+# =======================================================================================================================
+# =======================================================================================================================
+
+
+# =======================================================================================================================
+# ============ Losses ===================================================================================================
 
 def masked_bce_loss(y_true, y_pred_outputs):
     y_pred = y_pred_outputs
@@ -136,18 +158,22 @@ def masked_bce_loss(y_true, y_pred_outputs):
     mask = tf.squeeze(mask, axis=-1)  # Removes the last dimension if it's 1
     base_loss = tf.keras.losses.binary_crossentropy(y_true, y_pred, from_logits=False) 
     masked_loss = base_loss * mask
-    return tf.reduce_sum(masked_loss) / tf.reduce_sum(mask)
+
+    batch_size = tf.cast(tf.shape(y_true)[0], tf.float32)
+    return tf.reduce_sum(masked_loss) / tf.reduce_sum(mask) / batch_size # This might be kinda dumb, might be able to not use reduce_sum and avoid having to manually get batch_size
 
 def masked_mse_loss(y_true, y_pred_outputs):
     y_pred = y_pred_outputs
-    scale_factor = 1
+    scale_factor = 1 # For scaling the output and label pre-squaring the difference
 
     mask = tf.not_equal(y_true, -1.0)
     mask = tf.cast(mask, tf.float32)
     mask = tf.squeeze(mask, axis=-1)  # Removes the last dimension if it's 1
     base_loss = tf.keras.losses.mean_squared_error(scale_factor*y_true, scale_factor*y_pred)
     masked_loss = base_loss * mask
-    return tf.reduce_sum(masked_loss) / tf.reduce_sum(mask)
+
+    batch_size = tf.cast(tf.shape(y_true)[0], tf.float32)
+    return tf.reduce_sum(masked_loss) / tf.reduce_sum(mask) / batch_size
 
 def masked_mae_loss(y_true, y_pred_outputs):
     y_pred = y_pred_outputs
@@ -156,5 +182,10 @@ def masked_mae_loss(y_true, y_pred_outputs):
     mask = tf.cast(mask, tf.float32)
     mask = tf.squeeze(mask, axis=-1)  # Removes the last dimension if it's 1
     base_loss = tf.keras.losses.mean_absolute_error(y_true, y_pred)
-    masked_loss = (1 - base_loss) * mask
-    return tf.reduce_sum(masked_loss) / tf.reduce_sum(mask)
+    masked_loss = base_loss * mask
+
+    batch_size = tf.cast(tf.shape(y_true)[0], tf.float32)
+    return tf.reduce_sum(masked_loss) / tf.reduce_sum(mask) / batch_size
+
+# =======================================================================================================================
+# =======================================================================================================================
